% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma_numeric.R
\name{comma_numeric}
\alias{comma_numeric}
\title{Comma Numeric}
\usage{
comma_numeric(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A numeric vector
}
\description{
Takes a character vector of numbers with commas and returns a numeric vector
}
