% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hits.R
\name{get_hits}
\alias{get_hits}
\title{Get Hits}
\usage{
get_hits(year)
}
\arguments{
\item{year}{The year about which you want to know the jams}
}
\value{
A tibble of number one hits
}
\description{
Returns all number one hits for a given year (1959+)
}
\examples{
get_hits(1981)

}
