% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tri.R
\name{get_tri}
\alias{get_tri}
\title{Triangularize a Correlation Matrix}
\usage{
get_tri(cormat, lower = T)
}
\arguments{
\item{cormat}{Correlation Matrix}

\item{lower}{Lower or upper half (default = T) select F for upper half}
}
\value{
Matrix
}
\description{
Takes a correlation matrix and removes the redundant information
}
\examples{
cormat <- round(cor(mtcars), 2)
get_tri(cormat)

}
