% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_cormat.R
\name{reorder_cormat}
\alias{reorder_cormat}
\title{Reorder Correlation Matrix by correlation}
\usage{
reorder_cormat(cormat)
}
\arguments{
\item{cormat}{}
}
\value{
matrix
}
\description{
Reorder Correlation Matrix by correlation
}
\examples{
cormat <- round(cor(mtcars), 2)
reorder_cormat(cormat)
}
