% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_numeric.R
\name{split_numeric}
\alias{split_numeric}
\title{Split Numeric}
\usage{
split_numeric(x, split_char = ",")
}
\arguments{
\item{x}{A character vector}

\item{split_char}{The character to split by, as some regions use periods instead of commas (defaults to ",")}
}
\value{
A numeric vector
}
\description{
Takes a character vector of numbers with commas and returns a numeric vector
}
\examples{
comma_vec <- c("100,000", "250,000", "1,250", "4,242.42")
split_numeric(comma_vec)

period_vec <- c("100.000", "250.000", "1.250", "4.242,42")
split_numeric(period_vec, split_char = ".")

}
